smplify-x/requirements.txt
numpy>=1.16.2 #(1.23.5 worked for me)
# torch>=1.0.1.post2 
torchgeometry>=0.1.2
PyYAML==5.1.1
configargparse
smplx
pillow
opencv-python
tqdm
-e git+https://github.com/nghorbani/human_body_prior#egg=human_body_prior

